requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0 